[Feat] #41 - 마이페이지 레이아웃 구현 완료

+++ app/src/main/java/com/example/zupzup/domain/usecase/MakeReservationOnFireBaseUseCase.kt
@@ -8,5 +8,5 @@ interface MakeReservationOnFireBaseUseCase {
 
     operator fun invoke(
         reservationModel: ReservationModel
-    ): Flow<DataResult<Int>>
+    ): Flow<DataResult<Long>>
 }
